Change/typescript and yarn (#7)
* Changed from classic JS to TypeScript, and use Yarn instead of NPM.

* 2.1.0

* Typing for es6-promises.
