+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 example/node_modules
 build
 npm-debug.log
+yarn-error.log
