+++ .travis.yml
@@ -1,9 +1,3 @@
 language: node_js
-addons:
-  firefox: "46.0"
 node_js:
   - "7"
-  - "6"
-before_script:
-  - export DISPLAY=:99.0
-  - sh -e /etc/init.d/xvfb start
