+++ example/package.json
@@ -1,12 +1,6 @@
 {
   "name": "fitted-example",
   "scripts": {
-    "names": "babel-node ./names.js"
-  },
-  "dependencies": {
-    "babel-cli": "^6.6.5",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-stage-0": "^6.5.0"
+    "names": "node --harmony-async-await ./names.js"
   }
 }
