--- mocha.config.js
@@ -1,2 +1,0 @@
-require('babel-register');
-require('babel-polyfill');
